# Optimize byte stream split encoding