--- a/encoding/bytestreamsplit/bytestreamsplit_amd64.go
+++ b/encoding/bytestreamsplit/bytestreamsplit_amd64.go
@@ -19,7 +19,8 @@ var decodeFloatHasAVX2 = cpu.X86.HasAVX2
 
 var decodeDoubleHasAVX512 = cpu.X86.HasAVX512 &&
 	cpu.X86.HasAVX512F &&
-	cpu.X86.HasAVX512VL
+	cpu.X86.HasAVX512VL &&
+	cpu.X86.HasAVX512VBMI // VPERMB
 
 //go:noescape
 func encodeFloat(dst, src []byte)