--- a/encoding/test/test.go
+++ b/encoding/test/test.go
@@ -1,3 +1,5 @@
+//go:build go1.18
+
 package test
 
 import (