I was interested to learn about how to used the VPGATHER* and VPSCATTER* instructions in x86, so I used the BYTE_STREAM_SPLIT encoding as a test case.

Overall, the instructions bring measurable throughput improvements compared to the scalar versions of the algorithms, but they have high latency and don't seem to benefit from CPU pipelining. I tried to unroll the loops but it did not bring any throughput improvements. 

The gains are nevertheless interesting, the assembly code using scalar instructions was 20-30% faster than what the Go compiler generated, but using the vector instruction yields unlocked a lot more efficiencies:

```
name                             old time/op    new time/op     delta
Encode/BYTE_STREAM_SPLIT/float     13.7µs ± 0%      4.3µs ± 0%   -68.41%  (p=0.000 n=10+9)
Encode/BYTE_STREAM_SPLIT/double    35.7µs ± 0%      5.5µs ± 1%   -84.63%  (p=0.000 n=9+10)
Decode/BYTE_STREAM_SPLIT/float     12.8µs ± 0%      2.1µs ± 0%   -83.65%  (p=0.000 n=8+8)
Decode/BYTE_STREAM_SPLIT/double    28.0µs ± 0%      3.4µs ± 1%   -87.95%  (p=0.000 n=8+10)

name                             old speed      new speed       delta
Encode/BYTE_STREAM_SPLIT/float   2.91GB/s ± 0%   9.22GB/s ± 0%  +216.54%  (p=0.000 n=10+9)
Encode/BYTE_STREAM_SPLIT/double  2.24GB/s ± 0%  14.57GB/s ± 1%  +550.60%  (p=0.000 n=9+10)
Decode/BYTE_STREAM_SPLIT/float   3.13GB/s ± 0%  19.15GB/s ± 0%  +511.38%  (p=0.000 n=8+10)
Decode/BYTE_STREAM_SPLIT/double  2.86GB/s ± 0%  23.73GB/s ± 1%  +730.11%  (p=0.000 n=8+10)

name                             old value/s    new value/s     delta
Encode/BYTE_STREAM_SPLIT/float       728M ± 0%      2305M ± 0%  +216.54%  (p=0.000 n=10+9)
Encode/BYTE_STREAM_SPLIT/double      280M ± 0%      1822M ± 1%  +550.60%  (p=0.000 n=9+10)
Decode/BYTE_STREAM_SPLIT/float       783M ± 0%      4789M ± 0%  +511.37%  (p=0.000 n=8+10)
Decode/BYTE_STREAM_SPLIT/double      357M ± 0%      2966M ± 1%  +730.09%  (p=0.000 n=8+10)
```